Upstream mod to Minecraft 1.19.3
Update `worldToScreen` function in monkey way

Optimize imports

+++ src/main/java/me/qualterz/minecraft/lookaround/mixin/InGameHudMixin.java
@@ -1,5 +1,14 @@
 package me.qualterz.minecraft.lookaround.mixin;
 
+import me.qualterz.minecraft.lookaround.CameraState;
+import me.qualterz.minecraft.lookaround.LookaroundMod;
+import me.qualterz.minecraft.lookaround.ProjectionUtils;
+
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.hud.InGameHud;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.util.math.Vec3d;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -7,21 +16,12 @@ import org.spongepowered.asm.mixin.injection.ModifyArgs;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.hud.InGameHud;
-import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.util.math.Vec3d;
-
-import me.qualterz.minecraft.lookaround.CameraState;
-import me.qualterz.minecraft.lookaround.LookaroundMod;
-import me.qualterz.minecraft.lookaround.ProjectionUtils;
-
 @Mixin(InGameHud.class)
 public class InGameHudMixin {
     private CameraState camera;
 
-    private float offsetCrosshairX;
-    private float offsetCrosshairY;
+    private double offsetCrosshairX;
+    private double offsetCrosshairY;
 
     @Inject(method = "renderCrosshair", at = @At("HEAD"), cancellable = true)
     private void onRenderCrosshairBegin(MatrixStack matrices, CallbackInfo ci)
